Migrated to stable Quarkus version, moved to Panache, code cleanup. Moved logic out from controllers.

+++ src/main/java/io/quarkus/samples/petclinic/rest/PetTypesResource.java
@@ -6,18 +6,12 @@ import io.quarkus.samples.petclinic.service.ClinicService;
 
 import javax.annotation.security.RolesAllowed;
 import javax.inject.Inject;
-import javax.validation.Valid;
 import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
-import java.net.URI;
-import java.util.ArrayList;
 import java.util.Collection;
 
 @Path("api/pettypes")
@@ -41,7 +35,7 @@ public class PetTypesResource {
     @GET
     @Path("/{petTypeId}")
     @RolesAllowed({Roles.OWNER_ADMIN, Roles.VET_ADMIN})
-    public Response getPetType(@PathParam("petTypeId") int petTypeId){
+    public Response getPetType(@PathParam("petTypeId") long petTypeId){
         PetType petType = clinicService.findPetTypeById(petTypeId);
         if(petType == null){
             return Response.status(Response.Status.NOT_FOUND).build();
